# Conflicts:
#	src/main/resources/mapper/ElecRoundStuDao.xml

+++ src/main/java/com/server/edu/election/dto/StudentRePaymentDto.java
@@ -15,7 +15,7 @@ public class StudentRePaymentDto{
     private String courseCode;
     private String courseName;
     private Double credits;
-    private Double unitPrice;
+    private Integer unitPrice;
     private Double Payable;
     private Integer paid;
     private String studentCode;
@@ -104,11 +104,11 @@ public class StudentRePaymentDto{
         this.credits = credits;
     }
 
-    public Double getUnitPrice() {
+    public Integer getUnitPrice() {
         return unitPrice;
     }
 
-    public void setUnitPrice(Double unitPrice) {
+    public void setUnitPrice(Integer unitPrice) {
         this.unitPrice = unitPrice;
     }
